prometheus, Add ServiceMonitor CR
Deploy CNAO's ServiceMonitor on thr cluster, to serve components and
collect metrics according to the PrometheusRule instance.
Will only deploy if monitoring namespace is created, which is
configurable as CNAO container variable
The ServiceMonitor PrometheusRule instances are treated as a CNAO
component, and as such get the operator reconcile & labeling services.

Signed-off-by: Ram Lavi <ralavi@redhat.com>